app/backend/requirements.txt
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
python-multipart==0.0.20
psycopg2-binary==2.9.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
PyYAML==6.0.2
tenacity==8.5.0
pydantic-settings==2.4.0
email-validator==2.2.0

# Tinkoff Invest API SDK
tinkoff-investments==0.2.0b115
protobuf>=4.25.3,<5.0.0
grpcio>=1.56,<2

